Fix 32-bit integer overflow for Xtream IDs (issue #268)
Change all Xtream API ID fields from int to long to support IDs larger than
2,147,483,647. This includes stream_id, series_id, episode_id, category_id,
season_id, epg_id, and parent_id across all model classes.

- StreamInfo.StreamId/CategoryId: int → long
- Series.SeriesId/CategoryId: int → long
- SeriesInfo.CategoryId: int → long
- Episode.EpisodeId: int → long
- Category.CategoryId/ParentId: int → long
- Season.SeasonId: int → long
- EpgInfo.Id/EpgId: int → long
- SeriesStreamInfo.Episodes dict key: int → long
- IXtreamClient interface: all ID params int → long
- XtreamClient implementation: all ID params int → long
- StreamService: all ID params/methods int → long
- XtreamController route params: int → long
- XtreamVodProvider: int.Parse → long.Parse
- All channels (LiveTvService, VodChannel, SeriesChannel, CatchupChannel) updated
- ToGuid calls cast long→int for backward-compatible GUID encoding
- Configuration dictionaries keep int keys for XML serialization compat

+++ Jellyfin.Xtream/Api/XtreamController.cs
@@ -114,7 +114,7 @@ public class XtreamController(IXtreamClient xtreamClient) : ControllerBase
     /// <returns>An enumerable containing the streams.</returns>
     [Authorize(Policy = "RequiresElevation")]
     [HttpGet("LiveCategories/{categoryId}")]
-    public async Task<ActionResult<IEnumerable<StreamInfo>>> GetLiveStreams(int categoryId, CancellationToken cancellationToken)
+    public async Task<ActionResult<IEnumerable<StreamInfo>>> GetLiveStreams(long categoryId, CancellationToken cancellationToken)
     {
         Plugin plugin = Plugin.Instance;
         List<StreamInfo> streams = await xtreamClient.GetLiveStreamsByCategoryAsync(
@@ -146,7 +146,7 @@ public class XtreamController(IXtreamClient xtreamClient) : ControllerBase
     /// <returns>An enumerable containing the streams.</returns>
     [Authorize(Policy = "RequiresElevation")]
     [HttpGet("VodCategories/{categoryId}")]
-    public async Task<ActionResult<IEnumerable<StreamInfo>>> GetVodStreams(int categoryId, CancellationToken cancellationToken)
+    public async Task<ActionResult<IEnumerable<StreamInfo>>> GetVodStreams(long categoryId, CancellationToken cancellationToken)
     {
         Plugin plugin = Plugin.Instance;
         List<StreamInfo> streams = await xtreamClient.GetVodStreamsByCategoryAsync(
@@ -178,7 +178,7 @@ public class XtreamController(IXtreamClient xtreamClient) : ControllerBase
     /// <returns>An enumerable containing the streams.</returns>
     [Authorize(Policy = "RequiresElevation")]
     [HttpGet("SeriesCategories/{categoryId}")]
-    public async Task<ActionResult<IEnumerable<StreamInfo>>> GetSeriesStreams(int categoryId, CancellationToken cancellationToken)
+    public async Task<ActionResult<IEnumerable<StreamInfo>>> GetSeriesStreams(long categoryId, CancellationToken cancellationToken)
     {
         Plugin plugin = Plugin.Instance;
         List<Series> series = await xtreamClient.GetSeriesByCategoryAsync(
