+++ Jellyfin.Xtream/Client/IXtreamClient.cs
@@ -24,25 +24,25 @@ namespace Jellyfin.Xtream.Client;
 
 public interface IXtreamClient
 {
-    Task<EpgListings> GetEpgInfoAsync(ConnectionInfo connectionInfo, int streamId, CancellationToken cancellationToken);
+    Task<EpgListings> GetEpgInfoAsync(ConnectionInfo connectionInfo, long streamId, CancellationToken cancellationToken);
 
     Task<List<Category>> GetLiveCategoryAsync(ConnectionInfo connectionInfo, CancellationToken cancellationToken);
 
     Task<List<StreamInfo>> GetLiveStreamsAsync(ConnectionInfo connectionInfo, CancellationToken cancellationToken);
 
-    Task<List<StreamInfo>> GetLiveStreamsByCategoryAsync(ConnectionInfo connectionInfo, int categoryId, CancellationToken cancellationToken);
+    Task<List<StreamInfo>> GetLiveStreamsByCategoryAsync(ConnectionInfo connectionInfo, long categoryId, CancellationToken cancellationToken);
 
-    Task<List<Series>> GetSeriesByCategoryAsync(ConnectionInfo connectionInfo, int categoryId, CancellationToken cancellationToken);
+    Task<List<Series>> GetSeriesByCategoryAsync(ConnectionInfo connectionInfo, long categoryId, CancellationToken cancellationToken);
 
     Task<List<Category>> GetSeriesCategoryAsync(ConnectionInfo connectionInfo, CancellationToken cancellationToken);
 
-    Task<SeriesStreamInfo> GetSeriesStreamsBySeriesAsync(ConnectionInfo connectionInfo, int seriesId, CancellationToken cancellationToken);
+    Task<SeriesStreamInfo> GetSeriesStreamsBySeriesAsync(ConnectionInfo connectionInfo, long seriesId, CancellationToken cancellationToken);
 
     Task<PlayerApi> GetUserAndServerInfoAsync(ConnectionInfo connectionInfo, CancellationToken cancellationToken);
 
     Task<List<Category>> GetVodCategoryAsync(ConnectionInfo connectionInfo, CancellationToken cancellationToken);
 
-    Task<VodStreamInfo> GetVodInfoAsync(ConnectionInfo connectionInfo, int streamId, CancellationToken cancellationToken);
+    Task<VodStreamInfo> GetVodInfoAsync(ConnectionInfo connectionInfo, long streamId, CancellationToken cancellationToken);
 
-    Task<List<StreamInfo>> GetVodStreamsByCategoryAsync(ConnectionInfo connectionInfo, int categoryId, CancellationToken cancellationToken);
+    Task<List<StreamInfo>> GetVodStreamsByCategoryAsync(ConnectionInfo connectionInfo, long categoryId, CancellationToken cancellationToken);
 }
