+++ Jellyfin.Xtream/LiveTvService.cs
@@ -61,7 +61,7 @@ public class LiveTvService(IServerApplicationHost appHost, IHttpClientFactory ht
             ParsedName parsed = StreamService.ParseName(channel.Name);
             items.Add(new ChannelInfo()
             {
-                Id = StreamService.ToGuid(StreamService.LiveTvPrefix, channel.StreamId, 0, 0).ToString(),
+                Id = StreamService.ToGuid(StreamService.LiveTvPrefix, (int)channel.StreamId, 0, 0).ToString(),
                 Number = channel.Num.ToString(CultureInfo.InvariantCulture),
                 ImageUrl = channel.StreamIcon,
                 Name = parsed.Title,
@@ -179,7 +179,7 @@ public class LiveTvService(IServerApplicationHost appHost, IHttpClientFactory ht
                 {
                     items.Add(new()
                     {
-                        Id = StreamService.ToGuid(StreamService.EpgPrefix, streamId, epg.Id, 0).ToString(),
+                        Id = StreamService.ToGuid(StreamService.EpgPrefix, streamId, (int)epg.Id, 0).ToString(),
                         ChannelId = channelId,
                         StartDate = epg.Start,
                         EndDate = epg.End,
